seperate file for access control configuration
permissionsPolicy file will handle the access control configurations.
